solution/deploy/seeds/requirements.txt
sqlalchemy
pandas
python-dotenv